First functional prototype
The first slightly functional prototype of the search engine. As of right now, it only has .activeitem files indexed (and poorly at that). However, the search is capable of full regex.
From here, the main changes to be made are better indexing for more types of files, a dry run of the complete program (from scan to unpack to index to search), and a basic UI.